+++ example/lib/cache.dart
@@ -9,37 +9,32 @@ abstract class Cache implements BaseCache {
   static Cache get instance => _instance;
 
   @persistent
-  @CacheKey(path: 'device_id')
+  @Cached(path: 'device_id')
   CacheEntry<Iterable<String>> deviceId();
 
-  @CacheKey(fromJson: User.fromJson, toJson: userToJson)
+  @Cached(fromJson: User.fromJson, toJson: userToJson)
   CacheEntry<User> me();
 
-  @CacheKey(path: 'users')
+  @Cached(path: 'users')
   CacheEntry<List<User>> users();
 
   @MaxAge(Duration(seconds: 2))
-  @CacheKey(path: 'friends')
+  @Cached(path: 'all-friends')
   CacheEntry<int> friends();
 
-  @CacheKey(path: 'friends/{id}')
+  @Cached(path: 'friends/{id}')
   CacheEntry<String> friendById(
     @Path('id') int userId,
   );
 
-  @CacheKey(path: 'search/{query}')
-  CacheEntry<String> search(
-    @Path('query') String? query,
-  );
-
-  @CacheKey(path: 'likes/{date}')
+  @Cached(path: 'likes/{date}')
   CacheEntry<String> likes(
     @Path('date', convert: keyDateConvertor) DateTime date,
   );
 }
 
 dynamic userToJson(User user) {
-  return User(user.name, 58).toJson();
+  return user.toJson();
 }
 
 String keyDateConvertor(DateTime date) {